LayoutTests/imported/w3c/web-platform-tests/navigation-api/state/cross-document-location-api-expected.txt



FAIL entry.getState() behavior after cross-document location API navigation assert_equals: expected 1 but got 0
